LayoutTests/platform/mac-bigsur-wk2/http/tests/websocket/tests/hybi/workers/send-arraybuffer-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://127.0.0.1:8880/websocket/tests/hybi/workers/resources/check-binary-messages' failed: The operation couldn’t be completed. (kNWErrorDomainPOSIX error 57 - Socket is not connected)
WebSocket: Send ArrayBuffers in Web Workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS PASS: Message #0.
PASS PASS: Message #1.
PASS PASS: Message #2.
FAIL FAIL: closeEvent.wasClean should be true, but was: false
DONE
PASS successfullyParsed is true

TEST COMPLETE
